% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retention.R
\name{retention}
\alias{retention}
\title{Determining retention status for each student}
\usage{
retention(data, first_semester, next_semester, pidm_col = 1, term_code_col = 2)
}
\arguments{
\item{data}{Input data frame. This should have at least a column that uniquely
identifies the students and a column that indicates the term code that the
student was registered for.}

\item{first_semester}{The base semester for which retention is determined from.}

\item{next_semester}{The semester in which we are looking to see if the student
was retained.}

\item{pidm_col}{The index for the column that contains student identification
details. Default is 1.}

\item{term_code_col}{The index for the column that contains term code details.
Default is 2.}
}
\value{
A data frame of all students enrolled in \code{first_semester} with a
new column of logicals added indicating whether the student was retained to
\code{second_semester}.
}
\description{
\code{retention()} takes a data frame that includes at least an identification
number for each student and a term code for each student, and determines
whether a student who was enrolled in a certain semester was also enrolled in
another semester of the user's choice.
}
\examples{
df <- dplyr::tibble(pidm = c(1:5, 1, 2, 5, 1, 4),
        term_code = c(rep(202120,5), rep(202220,3), rep(202320,2)))
retention(df, 202120, 202220)
}
