% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mc.R
\name{scale_color_mc}
\alias{scale_color_mc}
\title{Use MC's colors for a color aesthetic}
\usage{
scale_color_mc(idx = 1:9)
}
\arguments{
\item{idx}{Integer vector of indices from the color palette to use.}
}
\description{
\code{scale_color_mc()} uses the main colors for the College as a color option for
a categorical data set.
}
\examples{
library(ggplot2)
ggplot(data = mtcars, aes(x = wt, y = mpg, color = factor(cyl))) +
        geom_point() +
        scale_color_mc()
}
