% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mc.R
\name{scale_fill_mc}
\alias{scale_fill_mc}
\title{Use MC's colors for a fill aesthetic}
\usage{
scale_fill_mc(idx = 1:9)
}
\arguments{
\item{idx}{Integer vector of indices from the color palette to use.}
}
\description{
\code{scale_fill_mc()} uses the main colors for the College as a fill option
for a categorical data set.
}
\examples{
ggplot(data = mtcars, aes(x = cyl, fill = factor(cyl))) + geom_histogram() + scale_fill_mc()
}
